SECURITY-863 LogAuditProvider allocates too much
`LogAuditProvider` always logs the `AuditEvent` even if audit logging
is disabled. This means the `AuditEvent` is always converted to a
`String` even if the trace log level is disabled. The profiled our
application and this causes a lot of allocations outside of TLABs for
us even though we have audit logging disabled.

 * don't log if the trace level is disabled

Issue: SECURITY-863
https://issues.jboss.org/browse/SECURITY-863